+++ changelog.md
@@ -10,4 +10,5 @@
 05.04.2025 - Successful configured PostgreSQL database on Hetzner VM\
 06.04.2025 - Manually deployed and run app on Hetzner VM\
 10.04.2025 - Reduced .jar size\
-10.04.2025 - Adds log to file with retention functionality
+10.04.2025 - Adds log to file with retention functionality\
+15.04.2025 - Adds incoming requests logging